Enable subscription to tags, modifier, or creator rooms.
By default the "room" is '*' which means all public tiddlers.

If the client emit a subscribe event, the data describes a thing to subscribe
to:

   tags/<some tag>
   modifier/<username>
   creator/<username>

This is not tested for unicode or weird chars and such yet.

To enable this, the listener now sends a json rep of the tiddler
down the beanstalk pipe, minus the _text_ field. Down the websocket
pipes is still just the uri.

Done this way to:

* keeping things constrained in a resty way
* avoid race conditions
* keep surface area low
* but maximize power (and extensibility) in the node server