Make presented links to tiddlers be friendly.
If we can detect that a tiddler is from a space, then go
ahead and present it with its friendly uri. Otherwise just
use the uri that's on the tiddler object.

The code that was there before seemed rather over complicated,
which suggests that other over complications lurk elsewhere.
Given time, a feel review would be useful.

@pads, if you could confirm this is kosher, feel free to deploy
it.